meeting-copilot-backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-community
langchain-core
langchain-huggingface
chromadb
psycopg2-binary
pgvector
python-dotenv
axios
transformers
torch
